+++ include/trifinger_cameras/pybullet_tricamera_driver.hpp
@@ -10,6 +10,7 @@
 
 #include <robot_interfaces/finger_types.hpp>
 #include <robot_interfaces/sensors/sensor_driver.hpp>
+#include <trifinger_cameras/camera_parameters.hpp>
 #include <trifinger_cameras/settings.hpp>
 #include <trifinger_cameras/tricamera_observation.hpp>
 
@@ -18,8 +19,8 @@ namespace trifinger_cameras
 /**
  * @brief Driver to get rendered camera images from pyBullet.
  */
-class PyBulletTriCameraDriver : public robot_interfaces::SensorDriver<
-                                    trifinger_cameras::TriCameraObservation>
+class PyBulletTriCameraDriver
+    : public robot_interfaces::SensorDriver<TriCameraObservation, TriCameraInfo>
 {
 public:
     PyBulletTriCameraDriver(
@@ -28,10 +29,20 @@ public:
         Settings settings = Settings());
 
     /**
+     * @brief Get the camera parameters (image size and calibration
+     * coefficients).
+     *
+     * **Important:**  The calibration coefficients are only set if the driver
+     * is initialized with a calibration file (see constructor).  Otherwise,
+     * they will be empty.
+     */
+    virtual TriCameraInfo get_sensor_info() override;
+
+    /**
      * @brief Get the latest observation from the three cameras
      * @return TricameraObservation
      */
-    trifinger_cameras::TriCameraObservation get_observation();
+    virtual trifinger_cameras::TriCameraObservation get_observation() override;
 
 private:
     //! @brief Python object to access cameras in pyBullet.
@@ -52,6 +63,9 @@ private:
 
     //! Number of robot time steps after which the next frame should be fetched.
     int frame_rate_in_robot_steps_;
+
+    //! Sensor info for the cameras.
+    TriCameraInfo sensor_info_ = {};
 };
 
 }  // namespace trifinger_cameras