+++ include/trifinger_cameras/pylon_driver.hpp
@@ -11,6 +11,8 @@
  */
 #pragma once
 
+#include <filesystem>
+
 #ifndef Pylon_FOUND
 #error Cannot use PylonDriver without Pylon.
 #endif
@@ -23,6 +25,7 @@
 
 #include <robot_interfaces/sensors/sensor_driver.hpp>
 #include <trifinger_cameras/camera_observation.hpp>
+#include <trifinger_cameras/camera_parameters.hpp>
 #include <trifinger_cameras/settings.hpp>
 
 namespace trifinger_cameras
@@ -43,15 +46,40 @@ void pylon_connect(std::string_view device_user_id,
  * @brief Driver for interacting with a camera via Pylon and storing
  * images using OpenCV.
  */
-class PylonDriver : public robot_interfaces::SensorDriver<CameraObservation>
+class PylonDriver
+    : public robot_interfaces::SensorDriver<CameraObservation, CameraInfo>
 {
 public:
     /**
-     * @param device_user_id "DeviceUserID" of the camera.
+     * @brief Connect to camera based on name.
+     *
+     * When using this constructor, the camera calibration coefficients returned
+     * by @ref get_sensor_info will be set to zero.
+     *
+     * @param device_user_id "DeviceUserID" of the camera.  Pass empty string to
+     * connect to first camera found (useful if only one camera is connected).
      * @param downsample_images If set to true (default), images are downsampled
      *     to half their original size.
+     * @param settings Settings for the camera.
      */
     PylonDriver(const std::string& device_user_id,
+                bool downsample_images = true,
+                Settings settings = Settings());
+
+    /**
+     * @brief Connect to camera based on calibration file.
+     *
+     * The provided calibration file is expected to be in YAML format and
+     * contain the "camera_name" (= DeviceUserID) as well as calibration
+     * coefficients.  The latter will be used in the CameraInfo returned by @ref
+     * get_sensor_info.
+     *
+     * @param camera_calibration_file Path to the camera calibration file.
+     * @param downsample_images If set to true (default), images are downsampled
+     *     to half their original size.
+     * @param settings Settings for the camera.
+     */
+    PylonDriver(const std::filesystem::path& camera_calibration_file,
                 bool downsample_images = true,
                 Settings settings = Settings());
 
@@ -69,19 +97,41 @@ public:
     static cv::Mat downsample_raw_image(const cv::Mat& image);
 
     /**
+     * @brief Get the camera parameters (image size and calibration
+     * coefficients).
+     *
+     * **Important:**  The calibration coefficients are only set if the driver
+     * is initialized with a calibration file (see constructor).  Otherwise,
+     * they will be empty.
+     */
+    virtual CameraInfo get_sensor_info() override;
+
+    /**
      * @brief Get the latest observation (image frame + timestamp of when the
      * frame's grabbed).
      * @return CameraObservation
      */
-    CameraObservation get_observation();
+    virtual CameraObservation get_observation() override;
 
 private:
     std::shared_ptr<const PylonDriverSettings> settings_;
+    trifinger_cameras::CameraInfo camera_info_ = {};
     std::string device_user_id_;
     const bool downsample_images_;
     Pylon::PylonAutoInitTerm auto_init_term_;
     Pylon::CInstantCamera camera_;
     Pylon::CImageFormatConverter format_converter_;
+
+    /**
+     * @brief Base constructor to be called by public constructors.
+     * @param downsample_images If set to true (default), images are downsampled
+     *     to half their original size.
+     * @param settings Settings for the camera.
+     */
+    PylonDriver(bool downsample_images, Settings settings);
+
+    //! Initialize the camera connection (to be called in the constructor).
+    void init(const std::string& device_user_id);
 
     void set_camera_configuration();
 };